requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.8b0
cffi==1.14.6
channels==3.0.4
click==8.0.1
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.7
django-crontab==0.7.1
django-filter==21.1
django-user-agents==0.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
isort==5.9.3
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
regex==2021.8.28
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
ua-parser==0.10.0
user-agents==2.2.0
zope.interface==5.4.0

django-debug-toolbar~=3.2.2